reorganize the codebase
