+++ packages/react-native-broadcast/src/Broadcast.ts
@@ -5,6 +5,7 @@ import {
   NativeEventEmitter,
   NativeModules,
 } from 'react-native';
+import { Presets } from './Presets';
 import {
   type CameraDirection,
   type MediaState,
@@ -14,31 +15,6 @@ import {
 const BroadcastEvents = new TypedNativeEventEmitter(
   new NativeEventEmitter(NativeModules.BroadcastEventEmitter),
 );
-
-export class Presets {
-  /**
-   * HD portrait mode preset. Recommended for most use cases.
-   */
-  static PORTRAIT_HD: Preset = {
-    width: 720,
-    height: 1280,
-    frameRate: 30,
-    videoBitrate: 3_000_000,
-    audioBitrate: 128_000,
-  };
-
-  /**
-   * Full HD portrait mode preset. Recommended for high-resolution broadcasts,
-   * but it can be slow on older devices.
-   */
-  static PORTRAIT_FULL_HD: Preset = {
-    width: 1080,
-    height: 1920,
-    frameRate: 30,
-    videoBitrate: 4_000_000,
-    audioBitrate: 128_000,
-  };
-}
 
 export class Broadcast {
   private readonly instanceId: string;
