+++ packages/react-native-broadcast/src/index.tsx
@@ -1,2 +1,3 @@
 export { Broadcast } from './Broadcast';
 export * from './BroadcastVideoView';
+export * from './Presets';
